+++ src/interfaces/index.ts
@@ -28,3 +28,4 @@ export * from './IPulseEmitter';
 export * from './IPlanRepository';
 export * from './IPlanDefinition';
 export * from './IPlanRepositoryStore';
+export * from './IPlanArchiver';